chore(rfs): rename rf to rfs as service name changed
